requirements.txt
Django==1.5.1
MySQL-python==1.2.4b4
distribute==0.6.28
dj-database-url==0.2.1
django-appconf==0.6
django-compressor==1.3
django-crispy-forms==1.2.5
django-debug-toolbar==0.9.4
six==1.3.0
wsgiref==0.1.2
